EBP-19: As a Solace Developer, I can use cache containers in integration testing.
Changed timeout to be 30 seconds instead of 30.0 seconds.

+++ test/testcontext/cache_util.go
@@ -274,6 +274,6 @@ func (ctx *testContextCommon) setupCache() error {
         if err != nil {
             return err
         }
-        err = ctx.waitForOperationalCache(time.Duration(30.0 * float64(time.Second)), CacheOperationalStateUp)
+        err = ctx.waitForOperationalCache(time.Duration(30 * time.Second), CacheOperationalStateUp)
         return err
 }
